requirements.txt
datasets
pillow
numpy
tqdm
torch
torchvision
torchaudio
torch-fidelity
lpips
